Fix requirements for Python 3.7 on Windows
Python 3.7 was released after tensorflow 1.12, so tensorflow 1.13
needs to be permitted; the use of "compatible" version matching is
usually done to permit anything with compatible versions to be chosen,
but the use of point release versions (~=1.12.0) was preventing that -
it clamped the version down to < 1.13.

Similarly with a number of other dependencies, which I got onto when
tracking installation failures of h5py, h5py now has binary wheels, but
newer versions are required (and obviously hdf5 must be installed to
enable those wheels). They are fragile, as they vendor the hdf5 dll,
and thus tables and h5py must match, but this seems to be primarily
a temporal concern.

Anyhow, I'll understand if I've gotten the wrong end of the tradeoff
here, and replace this patch with one that documents the tradeoffs
desired instead of changing them. Let me know.

With this patch + `vcpkg install hdf5[parallel]:x64-windows` and
installation of nvidia cuda 10 and related works, it appears as though
I have the base software installed; though the lack of a test suite
does make it somewhat hard to tell.