build and train should use config files
I'm not sure enough of the semantics yet - should there be a dataset
hdf5 file per class? - but it seems clear that we'll want configuration
for the ML stages as well, so this patch makes that config accessible
to that code, to allow future changes to leverage it.

I suspect that a map of classes and parameters are needed in the config
but that should flow naturally enough once this is usable.