Log to stderr
The goal was to fix a stray print to stdout which was breaking the
JSON output from classify.py (consumed by cacophony-processing).

Set things up so that logs go to stderr with a reasonable
format. stdout is then reserved for "data" output from the various
scripts. Most code sites that were using print are now using the
logging module.