Add files via upload
The class predator tracker takes mp4 file input from extract.py. 
The centroid of the animal is determined and a Kalman filter is used to smooth the
movement and reduce noise introduced by occlusion. 
A moving average filter is used to find the average velocity and this is used to 
calculate the future position of the animal at a given time.